Move inline content parser to beta API package
