+++ commonmark/src/main/java/org/commonmark/internal/InlineParserContextImpl.java
@@ -1,6 +1,6 @@
 package org.commonmark.internal;
 
-import org.commonmark.internal.inline.InlineContentParserFactory;
+import org.commonmark.parser.beta.InlineContentParserFactory;
 import org.commonmark.node.LinkReferenceDefinition;
 import org.commonmark.parser.InlineParserContext;
 import org.commonmark.parser.delimiter.DelimiterProcessor;
