+++ commonmark/src/main/java/org/commonmark/internal/inline/AutolinkInlineParser.java
@@ -3,8 +3,7 @@ package org.commonmark.internal.inline;
 import org.commonmark.node.Link;
 import org.commonmark.node.Text;
 import org.commonmark.parser.SourceLines;
-import org.commonmark.parser.beta.Position;
-import org.commonmark.parser.beta.Scanner;
+import org.commonmark.parser.beta.*;
 
 import java.util.Set;
 import java.util.regex.Pattern;
