+++ commonmark/src/main/java/org/commonmark/internal/inline/BackslashInlineParser.java
@@ -3,7 +3,7 @@ package org.commonmark.internal.inline;
 import org.commonmark.internal.util.Escaping;
 import org.commonmark.node.HardLineBreak;
 import org.commonmark.node.Text;
-import org.commonmark.parser.beta.Scanner;
+import org.commonmark.parser.beta.*;
 
 import java.util.Set;
 import java.util.regex.Pattern;
