+++ commonmark/src/main/java/org/commonmark/internal/inline/ParsedInlineImpl.java
@@ -1,13 +1,14 @@
 package org.commonmark.internal.inline;
 
 import org.commonmark.node.Node;
+import org.commonmark.parser.beta.ParsedInline;
 import org.commonmark.parser.beta.Position;
 
-public class ParsedInlineImpl extends ParsedInline {
+public class ParsedInlineImpl implements ParsedInline {
     private final Node node;
     private final Position position;
 
-    ParsedInlineImpl(Node node, Position position) {
+    public ParsedInlineImpl(Node node, Position position) {
         this.node = node;
         this.position = position;
     }
